[REFACTOR] 피드 업데이트 로직을 Application으로 이전
